+++ changelog.md
@@ -6,6 +6,7 @@
 * LLT-4967: Clear leftover STUN peer after meshnet is turned off
 * LLT-4913: Skip qos analytics ping for unconnected peers
 * LLT-4968: Make the default nurse initial_heartbeat_interval to be 5 minutes
+* LLT-4850: Add missing VPN meshnet address to analytics
 
 <br>